Fix settings for test
