--- src/components/molecules/ProfileHeader/index.test.js
@@ -1,15 +1,0 @@
-import React from 'react'
-import { shallow } from 'enzyme'
-import ProfileHeader from '.'
-
-const wrap = (props = {}) => shallow(<ProfileHeader {...props} />)
-
-it('renders children when passed in', () => {
-  const wrapper = wrap({ children: 'test' })
-  expect(wrapper.contains('test')).toBe(true)
-})
-
-it('renders props when passed in', () => {
-  const wrapper = wrap({ id: 'foo' })
-  expect(wrapper.find({ id: 'foo' })).toHaveLength(1)
-})
